% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{covf}
\alias{covf}
\title{Correlation estimators}
\usage{
covf(x, cor.method, scale.method, pda.method, lmin = NULL)
}
\arguments{
\item{x}{input matrix}

\item{cor.method}{"sd"  or "qn"}

\item{scale.method}{could be "pearson", "pair" and "gaussrank"}

\item{pda.method}{"nearpd" or FALSE}

\item{lmin}{min threshold of eigenvalues}
}
\value{
covmatrix, the estimated covariance matrix

cormatrix, the estimated correlation matrix

scale, the estimated scale vector

lmin, the min threshold of eigenvalues
}
\description{
Correlation estimators
}
\examples{
x = matrix(rnorm(100), ncol = 10)
paircorf(x)

}
