% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roblasso.R
\name{covlasso}
\alias{covlasso}
\title{Robust adaptive lasso based on robust correlation estimates}
\usage{
covlasso(
  x,
  y,
  cor.method = "pair",
  scale.method = "qn",
  pda.method = NULL,
  lmin = NULL,
  std = T,
  adaptive = T,
  cormatrix = NULL,
  scale = NULL
)
}
\arguments{
\item{x}{input design matrix}

\item{y}{input response vector}

\item{cor.method}{"sd" or "qn"}

\item{scale.method}{could be "pearson", "pair" and "gaussrank"}

\item{pda.method}{"nearpd" or FALSE}

\item{lmin}{min threshold of eigenvalues}

\item{std}{If TRUE the robust correlation matrix is used,
if FALSE the robust covariance matrix is used.}

\item{adaptive}{adaptive regularization penalties}

\item{cormatrix}{you could also use a correlation matrix as input}

\item{scale}{put in scales if you use cormatrix}
}
\value{
betahat_opt, the optimal estimation of beta obtained from this algorithm

lambda_opt is the optimal tuning parameter and sigma_opt is the optimal estimation of sigma.

The output also includes the estimated correlation matrix, the estimated covariance matrix and et cetera from the covf function.
}
\description{
Robust adaptive lasso based on robust correlation estimates
}
\examples{
dat = genevar()
y = dat$y
x = dat$x
fit = covlasso(x,y)
fit$betahat_opt

}
