% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{gaussrankf}
\alias{gaussrankf}
\title{Gaussrank cor}
\usage{
gaussrankf(x)
}
\arguments{
\item{x}{input matrix}
}
\value{

}
\description{
Gaussrank cor
}
\examples{
x = matrix(rnorm(100), ncol = 10)
gaussrankf(x)
}
