% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{genevar}
\alias{genevar}
\title{generating predictors and response randomly based on given parameters}
\usage{
genevar(n = 100, p = 20, e = 0, r = 0.5, gamma = 10)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{e}{contamination rate}

\item{r}{correlation coefficent}

\item{gamma}{magnitude of outliers}
}
\value{

}
\description{
generating predictors and response randomly based on given parameters
}
\examples{

}
