% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Grid_arrange_shared_legend}
\usage{
grid_arrange_shared_legend(
  ...,
  nrow = length(list(...)),
  ncol = 1,
  position = c("bottom", "right")
)
}
\arguments{
\item{position}{}
}
\description{
Grid_arrange_shared_legend
}
