% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nearpdf}
\alias{nearpdf}
\title{PD calibration}
\usage{
nearpdf(cormatrix, pda.method, lmin)
}
\arguments{
\item{cormatrix}{input cormatrix}

\item{pda.method}{"nearpd" or FALSE}

\item{lmin}{min threshold of eigenvalues}
}
\value{

}
\description{
PD calibration
}
\examples{
x = matrix(rnorm(100), ncol = 10)
nearpdf(cor(x), pda.method = "nearpd", lmin = 0.1)
}
