% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{paircorf}
\alias{paircorf}
\title{Pairwise correlation using Qn estimator}
\usage{
paircorf(x)
}
\arguments{
\item{x}{input matrix}
}
\value{

}
\description{
Pairwise correlation using Qn estimator
}
\examples{
x = matrix(rnorm(100), ncol = 10)
paircorf(x)
}
