% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{paircorxyf}
\alias{paircorxyf}
\title{Pairwise correlation between x and y}
\usage{
paircorxyf(x, y)
}
\arguments{
\item{x}{input vector}

\item{y}{input vector}
}
\value{
cor, the pairwise correlations between x and y
}
\description{
Pairwise correlation between x and y
}
\examples{
x = rnorm(100)
y = rnorm(100)
paircorxyf(x,y)
}
