% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roblasso.R
\name{covlasso}
\alias{covlasso}
\title{Robust adaptive lasso}
\usage{
covlasso(
  x,
  y,
  cor.method = "pair",
  scale.method = "qn",
  pda.method = NULL,
  lmin = NULL,
  std = T,
  adaptive = T,
  cormatrix = NULL,
  scale = NULL
)
}
\arguments{
\item{x}{input design matrix}

\item{y}{input response vector}

\item{cor.method}{"sd" or "qn"}

\item{scale.method}{could be "pearson", "pair" and "gaussrank"}

\item{pda.method}{"nearpd" or FALSE}

\item{lmin}{min threshold of eigenvalues}

\item{std}{If TRUE the robust correlation matrix is used,
if FALSE the robust covariance matrix is used.}

\item{adaptive}{adaptive regularization penalties}

\item{cormatrix}{you could also use a correlation matrix as input}

\item{scale}{put in scales if you use cormatrix}
}
\value{

}
\description{
Robust adaptive lasso
}
