% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{genevar}
\alias{genevar}
\title{generating predictors and response randomly using default settings}
\usage{
genevar(
  n = 100,
  p = 20,
  e = 0.05,
  r = 0.5,
  intercept = 0,
  beta = c(1, 2, 1, 2, 1, rep(0, p - 5)),
  gamma = 10,
  errorsigma = 1
)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{e}{contamination rate}

\item{r}{correlation coefficent}

\item{beta}{regression coefficients}

\item{gamma}{magnitude of outliers}
}
\value{
x, the generated design matrix

y, the generated response vector

beta, the regression coefficients we use
}
\description{
generating predictors and response randomly using default settings
}
