% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lminsel}
\alias{lminsel}
\title{Selection of the eigen value threshold}
\usage{
lminsel(cormatrix)
}
\arguments{
\item{cormatrix}{input cormatrix}
}
\value{

}
\description{
Selection of the eigen value threshold
}
